🐛 accept decimal period, pulse and offset for timing clock (#2714)
https://github.com/plantuml/plantuml/issues/2711

The CommandClock regex only matched integer values for the clock's
period, pulse and offset, so

    clock "MyClock" as Clk with period 12.5 offset 0.1

failed to parse. The downstream rendering code (PanelsClock, Panels.xOfTime,
drawHorizontalBetweenTimes) already operates on doubles, so the values were
being constrained needlessly at the parser level.

Allow decimals in the regex and propagate the values as double through
PlayerClock, PanelsClock and TimingDiagram.createPlayerClock /
getClockValue. BigDecimal.valueOf(double) is used in place of
new BigDecimal(double) to avoid binary-precision noise.

Adds a non-regression test reproducing the original bug report.

Closes #2711

+++ src/main/java/net/sourceforge/plantuml/timingdiagram/PlayerClock.java
@@ -34,6 +34,8 @@
  */
 package net.sourceforge.plantuml.timingdiagram;
 
+import java.math.BigDecimal;
+
 import net.sourceforge.plantuml.klimt.color.Colors;
 import net.sourceforge.plantuml.skin.ArrowConfiguration;
 import net.sourceforge.plantuml.stereo.Stereotype;
@@ -44,12 +46,12 @@ import net.sourceforge.plantuml.timingdiagram.graphic.PanelsClock;
 
 public class PlayerClock extends Player {
 
-	private final int period;
-	private final int pulse;
-	private final int offset;
+	private final BigDecimal period;
+	private final BigDecimal pulse;
+	private final BigDecimal offset;
 
-	public PlayerClock(String title, ISkinParam skinParam, TimingRuler ruler, int period, int pulse, int offset,
-			boolean compact, Stereotype stereotype) {
+	public PlayerClock(String title, ISkinParam skinParam, TimingRuler ruler, BigDecimal period, BigDecimal pulse,
+			BigDecimal offset, boolean compact, Stereotype stereotype) {
 		super(title, skinParam, ruler, compact, stereotype, null, SName.clock, 30);
 		this.period = period;
 		this.pulse = pulse;
@@ -78,7 +80,7 @@ public class PlayerClock extends Player {
 		throw new UnsupportedOperationException();
 	}
 
-	public final int getPeriod() {
+	public final BigDecimal getPeriod() {
 		return period;
 	}
